+++ mobile/android/android-components/components/service/firefox-accounts/src/main/java/mozilla/components/service/fxa/sync/SyncManager.kt
@@ -6,6 +6,7 @@ package mozilla.components.service.fxa.sync
 
 import mozilla.components.concept.sync.SyncableStore
 import mozilla.components.service.fxa.SyncConfig
+import mozilla.components.service.fxa.SyncEngine
 import mozilla.components.support.base.log.logger.Logger
 import mozilla.components.support.base.observer.Observable
 import mozilla.components.support.base.observer.ObserverRegistry
@@ -136,7 +137,7 @@ abstract class SyncManager(
      */
     internal fun start() = synchronized(this) {
         logger.debug("Enabling...")
-        syncDispatcher = initDispatcher(newDispatcher(syncDispatcher, syncConfig.syncableStores))
+        syncDispatcher = initDispatcher(newDispatcher(syncDispatcher, syncConfig.supportedEngines))
         logger.debug("set and initialized new dispatcher: $syncDispatcher")
     }
 
@@ -151,13 +152,13 @@ abstract class SyncManager(
         syncDispatcher = null
     }
 
-    internal abstract fun createDispatcher(stores: Set<String>): SyncDispatcher
+    internal abstract fun createDispatcher(supportedEngines: Set<SyncEngine>): SyncDispatcher
 
     internal abstract fun dispatcherUpdated(dispatcher: SyncDispatcher)
 
     private fun newDispatcher(
         currentDispatcher: SyncDispatcher?,
-        stores: Set<String>
+        supportedEngines: Set<SyncEngine>
     ): SyncDispatcher {
         // Let the existing dispatcher, if present, cleanup.
         currentDispatcher?.close()
@@ -166,7 +167,7 @@ abstract class SyncManager(
         currentDispatcher?.unregister(dispatcherStatusObserver)
 
         // Create a new dispatcher bound to current stores and account.
-        return createDispatcher(stores)
+        return createDispatcher(supportedEngines)
     }
 
     private fun initDispatcher(dispatcher: SyncDispatcher): SyncDispatcher {