+++ mobile/android/android-components/docs/changelog.md
@@ -44,6 +44,9 @@ permalink: /changelog/
   * **browser-engine-servo**
     * ❌ We removed the `browser-engine-servo` component since it was not being maintained, updated and used.
 
+* **service-firefox-accounts**
+  * ⚠️ **This is a breaking change**: `SyncConfig`'s `syncableStores` has been renamed to `supportedEngines`, expressed via new enum type `SyncEngine`.
+
 # 11.0.0
 
 * [Commits](https://github.com/mozilla-mobile/android-components/compare/v10.0.0...v11.0.0)